[IMP] Added tests_rest_alerts
